Refactor dimensions and tabulations logic into separate module
so that we can add tests later, when we have time.
