+++ lib/reports/ReportCollisionDirectory.js
@@ -65,10 +65,6 @@ class ReportCollisionDirectory extends ReportBaseCrash {
   }
 
   transformData(location, rawData) {
-    /*
-     * TODO: this violates the "no shared references" invariant in `ReportBase#transformData`.
-     * We should address that eventually, but there isn't time right now.
-     */
     const collisions = rawData.collisions.map(
       ReportCollisionDirectory.getCollisionDirectoryEntry,
     );
