+++ lib/reports/ReportCollisionTabulation.js
@@ -2,178 +2,16 @@
 import ArrayUtils from '@/lib/ArrayUtils';
 import { ReportBlock, ReportType, SortDirection } from '@/lib/Constants';
 import ReportBaseCrash from '@/lib/reports/ReportBaseCrash';
+import {
+  CollisionDimension,
+  EventCrossTabulation,
+  InvolvedCrossTabulation,
+} from '@/lib/reports/tabulation/CollisionTabulation';
 import TimeFormatters from '@/lib/time/TimeFormatters';
-
-class CollisionDimension {
-  constructor({
-    description,
-    entries,
-    event,
-    fn,
-    nullValue = null,
-    otherDescriptionFn = values => `${values.length} other values`,
-  }) {
-    this.description = description;
-    this.entries = entries;
-    this.event = event;
-    this.fn = fn;
-    this.nullValue = nullValue;
-    if (!this.entries.has(this.nullValue)) {
-      this.entries.set(this.nullValue, { description: 'Unknown' });
-    }
-    this.otherDescriptionFn = otherDescriptionFn;
-  }
-}
-
-class CrossTabulation {
-  constructor(dimX, dimY) {
-    this.dimX = dimX;
-    this.dimY = dimY;
-    this.table = new Map();
-    this.total = 0;
-    this.dimY.entries.forEach((valueY, y) => {
-      const values = new Map();
-      this.dimX.entries.forEach((valueX, x) => {
-        values.set(x, 0);
-      });
-      this.table.set(y, { total: 0, values });
-    });
-  }
-
-  incr(x0, y0) {
-    let y = y0;
-    if (y === null || !this.table.has(y)) {
-      y = this.dimY.nullValue;
-    }
-    const tableY = this.table.get(y);
-
-    let x = x0;
-    if (x === null || !tableY.values.has(x)) {
-      x = this.dimX.nullValue;
-    }
-    const valueX = tableY.values.get(x);
-    this.total += 1;
-    tableY.total += 1;
-    tableY.values.set(x, valueX + 1);
-  }
-
-  getOrderX() {
-    const { entries, nullValue } = this.dimX;
-    const keys = Array.from(entries.keys());
-    const keysNoNull = keys.filter(x => x !== nullValue);
-    const orderX = ArrayUtils.sortBy(keysNoNull, x => x);
-    orderX.push(nullValue);
-    return orderX;
-  }
-
-  getOrderY({
-    hideNull = false,
-    hideOther = false,
-    limit = 10,
-    sortByTotal = true,
-    sortDirection = SortDirection.DESC,
-  }) {
-    const { entries, nullValue } = this.dimY;
-    const keys = Array.from(entries.keys());
-    const keysNoNull = keys.filter(y => y !== nullValue);
-    let sortKey;
-    if (sortByTotal) {
-      sortKey = y => this.table.get(y).total;
-    } else {
-      sortKey = y => y;
-    }
-    const orderY = ArrayUtils.sortBy(keysNoNull, sortKey, sortDirection);
-    const n = hideNull ? keysNoNull.length : keys.length;
-    if (n <= limit) {
-      if (!hideNull) {
-        orderY.push(nullValue);
-      }
-      return orderY;
-    }
-    const hasNull = this.table.get(nullValue).total > 0;
-    if (!hideNull && hasNull) {
-      const orderYRest = orderY.slice(limit - 2);
-      const orderYLimit = orderY.slice(0, limit - 2);
-      orderYLimit.push(nullValue);
-      if (!hideOther) {
-        orderYLimit.push(orderYRest);
-      }
-      return orderYLimit;
-    }
-    const orderYRest = orderY.slice(limit - 1);
-    const orderYLimit = orderY.slice(0, limit - 1);
-    if (!hideOther) {
-      orderYLimit.push(orderYRest);
-    }
-    return orderYLimit;
-  }
-
-  getOrderedTable(orderX, orderY) {
-    const nx = orderX.length;
-    const ny = orderY.length;
-    const orderedTable = orderY.map((y) => {
-      const row = new Array(nx + 1).fill(0);
-      const ys = Array.isArray(y) ? y : [y];
-      ys.forEach((yy) => {
-        const { total, values } = this.table.get(yy);
-        orderX.forEach((x, j) => {
-          row[j] += values.get(x);
-        });
-        row[nx] += total;
-      });
-      return row;
-    });
-    const rowTotals = new Array(nx + 1).fill(0);
-    for (let i = 0; i < ny; i++) {
-      for (let j = 0; j <= nx; j++) {
-        rowTotals[j] += orderedTable[i][j];
-      }
-    }
-    orderedTable.push(rowTotals);
-    return orderedTable;
-  }
-}
-
-class EventCrossTabulation extends CrossTabulation {
-  constructor(dimX, dimY) {
-    super(dimX, dimY);
-    if (!dimX.event || !dimY.event) {
-      throw new Error('expected event-level dimensions!');
-    }
-  }
-
-  tabulate(collisions) {
-    collisions.forEach((event) => {
-      const x = this.dimX.fn(event);
-      const y = this.dimY.fn(event);
-      this.incr(x, y);
-    });
-  }
-}
-
-class InvolvedCrossTabulation extends CrossTabulation {
-  constructor(dimX, dimY) {
-    super(dimX, dimY);
-    if (dimX.event || dimY.event) {
-      throw new Error('expected involved-level dimensions!');
-    }
-  }
-
-  tabulate(collisions) {
-    collisions.forEach(({ involved, ...event }) => {
-      involved.forEach((person) => {
-        const x = this.dimX.fn(event, person);
-        const y = this.dimY.fn(event, person);
-        this.incr(x, y);
-      });
-    });
-  }
-}
 
 class ReportCollisionTabulation extends ReportBaseCrash {
   constructor() {
     super();
-    // TODO: proper initialization system for report modules
     this.initedDims = false;
   }
 
@@ -349,6 +187,10 @@ class ReportCollisionTabulation extends ReportBaseCrash {
 
       orderedTable[i].forEach((value, j) => {
         if (j >= nx) {
+          /*
+           * Each `orderedTable[i]` has an element at index `nx` containing the total across
+           * all `dimX` values.  We skip that here.
+           */
           return;
         }
